Flytter validering til util klasse, fjernet gammel test som greide å duplisere seg
